feat: implement modular API routes and consolidate server utility functions
